refactor: move pkg/units from orchestrator to shared
pkg/units holds byte-size helpers with no orchestrator-specific
dependencies. Moving it to shared lets shared packages (e.g. storage)
reuse them without importing orchestrator.
